# docker_images/csharp/wrapper/src/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.",
    "version" : "1.0.0",
    "title" : "Azure IOT End-to-End Test Wrapper Rest Api"
  },
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/device/connect/{transportType}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Connect to the azure IoT Hub as a device",
        "operationId" : "Device_Connect",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "caCertificate",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/certificate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/device/{connectionId}/disconnect" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Disconnect the device",
        "operationId" : "Device_Disconnect",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/createFromConnectionString/{transportType}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Create a device client from a connection string",
        "operationId" : "Device_CreateFromConnectionString",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "caCertificate",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/certificate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/device/createFromX509/{transportType}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Create a device client from X509 credentials",
        "operationId" : "Device_CreateFromX509",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "in" : "body",
          "name" : "X509",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/device/{connectionId}/connect2" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Connect the device",
        "operationId" : "Device_Connect2",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/reconnect" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Reconnect the device",
        "operationId" : "Device_Reconnect",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "forceRenewPassword",
          "in" : "query",
          "description" : "True to force SAS renewal",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/disconnect2" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Disconnect the device",
        "operationId" : "Device_Disconnect2",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/destroy" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Disconnect and destroy the device client",
        "operationId" : "Device_Destroy",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/enableMethods" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Enable methods",
        "operationId" : "Device_EnableMethods",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/waitForMethodAndReturnResponse/{methodName}" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Wait for a method call, verify the request, and return the response.",
        "description" : "This is a workaround to deal with SDKs that only have method call operations that are sync.  This function responds to the method with the payload of this function, and then returns the method parameters.  Real-world implemenatations would never do this, but this is the only same way to write our test code right now (because the method handlers for C, Java, and probably Python all return the method response instead of supporting an async method call)",
        "operationId" : "Device_WaitForMethodAndReturnResponse",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "methodName",
          "in" : "path",
          "description" : "name of the method to handle",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "requestAndResponse",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodRequestAndResponse"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/enableC2dMessages" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Enable c2d messages",
        "operationId" : "Device_EnableC2dMessages",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/event" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Send an event",
        "operationId" : "Device_SendEvent",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "eventBody",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/eventBody"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/c2dMessage" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "Wait for a c2d message",
        "operationId" : "Device_WaitForC2dMessage",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/eventBody"
            }
          }
        }
      }
    },
    "/device/{connectionId}/enableTwin" : {
      "put" : {
        "tags" : [ "device" ],
        "summary" : "Enable device twins",
        "operationId" : "Device_EnableTwin",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/twin" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "Get the device twin",
        "operationId" : "Device_GetTwin",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/twin"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "device" ],
        "summary" : "Updates the device twin",
        "operationId" : "Device_PatchTwin",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "twin",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/twin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/device/{connectionId}/twinDesiredPropPatch" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "Wait for the next desired property patch",
        "operationId" : "Device_WaitForDesiredPropertiesPatch",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/twin"
            }
          }
        }
      }
    },
    "/device/{connectionId}/connectionStatus" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "get the current connection status",
        "operationId" : "Device_GetConnectionStatus",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/device/{connectionId}/connectionStatusChange" : {
      "get" : {
        "tags" : [ "device" ],
        "summary" : "wait for the current connection status to change and return the changed status",
        "operationId" : "Device_WaitForConnectionStatusChange",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "connectResponse" : {
      "type" : "object",
      "properties" : {
        "connectionId" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "result of a connection to a service, device, or module client",
      "example" : {
        "connectionId" : "connectionId"
      }
    },
...
    "certificate" : {
      "type" : "object",
      "properties" : {
        "cert" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "certificate in the body of a message"
    },
    "eventBody" : {
      "type" : "object",
      "properties" : {
        "body" : {
          "type" : "object",
          "description" : "payload to send to the method",
          "properties" : { }
        },
        "horton_flags" : {
          "type" : "object",
          "description" : "flags used by horton",
          "properties" : { }
        },
        "attributes" : {
          "type" : "object",
          "description" : "Message attributes",
          "properties" : { }
        }
      },
      "description" : "body for an invoming or outgoing event or message",
      "example" : {
        "horton_flags" : "{}",
        "attributes" : "{}",
        "body" : "{}"
      }
    },
    "methodRequestAndResponse" : {
      "type" : "object",
      "properties" : {
        "requestPayload" : {
          "type" : "object",
          "description" : "payload for the request that arrived from the service.  Used to verify that the correct request arrived.",
          "properties" : { }
        },
        "responsePayload" : {
          "type" : "object",
          "description" : "payload for the response to return to the service.  Used to verify that the correct request arrived.",
          "properties" : { }
        },
        "statusCode" : {
          "type" : "integer",
          "description" : "status code to return to the service"
        }
      },
      "description" : "parameters and response for a sync method call"
    },
...
  }
}